+++ packages/adminui-vue/CHANGELOG.md
@@ -10,9 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
    pressing a "Refresh" in UTableEntity will refresh a table data AND lookups caches for entities used by current table.
    Default is 0 - do not refresh lookups. *WARNING* - entities with `Session` and `SessionEntity` cache type will not be refreshed.
  - `UTableEntity` filters: for columns of type Entity and Many added the ability to set a repository for a dropdown list (dropdown filter)
- - `UTableEntity` ability to register column definition globally and use this definition through the
-   `customSettings.UTableEntityColumnType` of the attribute. This is useful to define a template for
-   column cell slot globally for example
+ - `UTableEntity` ability to register custom cell templates for columns globally and use them through the
+   `customSettings.columnTemplate` of the column attribute
 
 ### Changed
  - `UTableEntity`: lookups are loaded in parallel to speed-up table loading